+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallEvent.swift
@@ -35,10 +35,12 @@ enum InternalSuperwallEvent {
   struct AppInstall: TrackableSuperwallEvent {
     let superwallEvent: SuperwallEvent = .appInstall
     let appInstalledAtString: String
+    let hasPurchaseController: Bool
     var customParameters: [String: Any] = [:]
     func getSuperwallParameters() async -> [String: Any] {
       return [
-        "application_installed_at": appInstalledAtString
+        "application_installed_at": appInstalledAtString,
+        "using_purchase_controller": hasPurchaseController
       ]
     }
   }